Feat: Adds i18n to error messages in player screen
Implements internationalization for error messages displayed in the player screen.

This change ensures that error messages are displayed in the user's preferred language (Korean or English) by checking the language setting in the HiveManager.

The tests have been updated to reflect the changes in error message localization and to add korean test cases.