fix: Improves temporary PDF file cleanup process
Refactors the temporary PDF file cleanup process to be asynchronous for better performance and prevents blocking of the main thread.

Also improves the robustness of the cleanup process by:
- Using `getTemporaryDirectory` to determine the temporary directory path
- Adding checks to ensure that the temporary directory exists before attempting to list files
- Adding more logging to help diagnose cleanup failures

It also waits shortly after disposing of the resources before attempting to clean up the files in order to allow file handles to be completely released.