feat: Adds NavigatorObserver for loading service
Implements a NavigatorObserver to track route changes and notify the LectureLoadingService.

This allows the loading service to react to navigation events, such as collapsing the loading bar to a bubble when navigating away from the home screen during an ongoing lecture creation.

Also, it ensures the loading bar is hidden when a lecture creation is completed and the user navigates away from the home screen, with a 15-second timeout.

The route settings are now consistently passed to the MaterialPageRoute constructor to propagate route information, enabling the NavigatorObserver to function correctly.